add duration settings
