+++ FullScreenTranslator/Helpers/Translator.swift
